[OSDEV-2542] Split DataPartnersFilter module files
Refactor DataPartnersFilter into a dedicated component folder with separate constants and styles modules, and switch consumers to import the component file directly. This also includes related filter/sidebar cleanup to keep the Spotlight search flow compiling and stable.
