+++ src/react/src/components/Filters/DataPartnersFilter/styles.js
@@ -1,0 +1,8 @@
+const makeDataPartnersFilterStyles = () =>
+    Object.freeze({
+        sharedFacilitiesCheckbox: Object.freeze({
+            marginLeft: '16px',
+        }),
+    });
+
+export default makeDataPartnersFilterStyles;
