+++ src/react/src/util/styles.js
@@ -287,13 +287,6 @@ export const makeCustomDropdownIndicatorStyles = Object.freeze({
     }),
 });
 
-export const makeDataPartnersFilterStyles = () =>
-    Object.freeze({
-        sharedFacilitiesCheckbox: Object.freeze({
-            marginLeft: '16px',
-        }),
-    });
-
 export const makeContributeProductionLocationStyles = theme =>
     Object.freeze({
         mainContainerStyles: Object.freeze({
